Close #32 Add coverage w/ pytest-cov
- Update .gitignore to ignore .coverage folder
- Move logic into src/cunao.py since we need to pass a module as args to
pytest-cov
- Refactor cunao.py file to import and use src/cunao.py, keeping
backwards retrocompatibility
